sbin/attest-enroll: Sign assets
We're really going to want to be able to authenticate enrollment
assets sent back to the attestation client upon success.

We need this to prevent impersonation attacks where an attacker that
knows a client's EKpub substitutes their own attestation response.
Rather than analyze the potential impact of such an attack, and develop
defenses on a per-asset-type basis, let us sign all the artifacts and
validate on the client side.

The problem is that on install, if we're booting untrusted media (e.g.,
PXE), then we can't obtain a trust anchor for validating the signatures.
The only options to deal with that are: TOFU (the winning PXE server
gives the client an anchor) or install from trusted media (e.g., a USB
flashdrive).