hcp: label containers, not container images
Labels are used to "find" containers corresponding to an instance of
interest. If they are baked into images at build-time they lose their
uselessness if the images are tagged with different prefix/suffix naming
as a way of distinguishing instances. Instead, we defer all labels to
run-time, ensuring they are assigned according to the intended
instance/naming.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>