hcp/run: respect NO_REINIT for container init/start dependencies
Until this commit, NO_REINIT was to relax dependencies on the need to
rebuild source code (when images are out of date) before initializing
and/or starting services. Now, it controls whether the dependency of
container-start on state-initialization is an order-only or regular. (I.e.
with NO_REINIT, the dependency is only on service-state _existing_, not on
it being prior to container-start.)

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>